Added StatView and Stat View Model
Files Added
